removed all but the one successfully case of basic streaming STT
