--- test_01.py
@@ -1,2 +1,0 @@
-import torchaudio
-wav, sr = torchaudio.load('soundsample.wav')
